:art: Restructure and format
Restructure auth code, add comments for all functions, and slightly alter formatting.
